backend/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.16.0
et_xmlfile==2.0.0
fonttools==4.56.0
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.1
msal==1.32.0
numpy==2.2.3
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pycparser==2.22
PyJWT==2.10.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
